Added a new method called "runInputLoop"
runInputLoop(number_of_menu_options, menu_options) is used in any menu method, and will collect user_input and ensure that it is sanitized to match one of the existing options for any given menu.

You must include the number of menu options for that method, and a String array with all of the available options. These are both to be declared at the top of the menu method in question. 

For example...

int number_of_menu_options = 4;
		String [] menu_options = new String[number_of_menu_options];
		String user_input;
		
		//Program option 1
		menu_options[0] = "yes";
		menu_options[1] = "1";
		
		//Program option 2
		menu_options[2] = "no";
		menu_options[3] = "2";

This should be done in every menu method. 


Additionally