runTraining() Update
runTraining was updated such that it is no longer a dead end pathway. It now has the functionality to choose through 5 different menu options. It can choose any of the four levels, or return to the main menu. 

Note that all of the options currently return to the main menu, until the case statements are updated to send them to the proper levels.